# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Crimson Circle Energy Company",
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http:/crimsonce.site/api/v1"
        }
    ],
    "paths": {
...
        "/network/add": {
            "post": {
                "tags": [
                    "Network"
                ],
                "description": "Add professional to network",
                "operationId": "addToNetwork",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{contact->FirstName} {contact->LastName} has been added to your network",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marilou Schaden Quitzon has been added to your network"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "contact_id": {
                                                        "type": "number",
                                                        "example": "1"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/network/delete": {
            "delete": {
                "tags": [
                    "Network"
                ],
                "description": "Remove professional to network",
                "operationId": "deleteToNetwork",
                "parameters": [
                    {
                        "name": "email",
                        "in": "path",
                        "description": "User email",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{contact->FirstName} {contact->LastName} has been removed from your network",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Marilou Schaden Quitzon has been removed from your network"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/network/add-random/{quantity}": {
            "get": {
                "tags": [
                    "Network"
                ],
                "description": "Add random number of contacts to network",
                "operationId": "addRandomContacts",
                "parameters": [
                    {
                        "name": "quantity",
                        "in": "path",
                        "description": "Number of users to add",
                        "required": true,
                        "schema": {
                            "type": "number"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{no_contacts_added} contacts have been added to your network",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "result": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "10 contacts have been added to your network"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}